+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Fix a false positive for `RSpec/PendingWithoutReason` when pending/skip has a reason inside an example group. ([@ydah])
 - Add support for `RSpec/ContainExactly` when calling `match_array` with no arguments. ([@ydah])
 - Fix an incorrect autocorrect for `RSpec/MatchArray` when calling `match_array` with an empty array literal. ([@bquorning])
+- Fix an incorrect autocorrect for `RSpec/Rails/TravelAround` when other before block exist. ([@ydah])
 
 ## 2.19.0 (2023-03-06)